services/route-engine/pyproject.toml
[project]
name = "route-engine"
version = "0.1.0"
description = "SMWCS Route Optimisation Engine"
authors = [
    {name = "dee-mee",email = "derekmuriuki2@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "confluent-kafka (>=2.13.2,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ortools (>=9.15.6755,<10.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "prometheus-client (>=0.24.1,<0.25.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "tenacity (>=9.1.4,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ruff (>=0.15.7,<0.16.0)"
]
